+++ core/src/main/java/uk/gov/justice/services/core/featurecontrol/remote/NonCachingFeatureProvider.java
@@ -1,6 +1,5 @@
 package uk.gov.justice.services.core.featurecontrol.remote;
 
-import uk.gov.justice.services.core.featurecontrol.FeatureFetcher;
 import uk.gov.justice.services.core.featurecontrol.domain.Feature;
 
 import java.util.Optional;
@@ -10,10 +9,10 @@ import javax.inject.Inject;
 public class NonCachingFeatureProvider {
 
     @Inject
-    private FeatureFetcher featureFetcher;
+    private FeatureFetcherFacade featureFetcherFacade;
 
     public Optional<Feature> lookup(final String featureName) {
-        return featureFetcher.fetchFeatures()
+        return featureFetcherFacade.fetchFeatures()
                 .stream()
                 .filter(feature -> featureName.equals(feature.getFeatureName()))
                 .findFirst();